Retain formatted email address in sendouts
The interceptor calls 'mail.to' to retrieve the email address. However, the mail gem will only return the address, and not the full formatted string. We can call `.addrs` (https://github.com/mikel/mail/blob/master/lib/mail/fields/common_address_field.rb#L47) but there's no easy way to access it. By removing all of this, we don't mangle the to/cc/bcc fields, and instead pass it straight through to Sendgrid which will handle it correctly.

This also fixes a bug where bcc and cc aren't treated like they say, and are converted to 'To' addresses.

This also removes the 'dummy_recipient' which has shown to be flagged as spam by GMail.